Fix encoding issues with open()
Traceback (most recent call last):
  File "setup.py", line 7, in <module>
    readme = f.read()
  File "/usr/lib/python3.5/encodings/ascii.py", line 26, in decode
    return codecs.ascii_decode(input, self.errors)[0]
UnicodeDecodeError: 'ascii' codec can't decode byte 0xe2 in position 4610: ordinal not in range(128)

When opening files in text mode, a good practice is to use io.open()
with an explicit encoding argument. io.open() works in Python 2.6 and
all later versions. In Python 3, io.open() is an alias for the built-in
open().