exclude cache from memory report as suggested in #47
Arguably, cache space is also reclaimable, meaning that, under pressure,
it can be reclaimed instead of triggering an OOM. Wether to substract it
or not would be a debate in itself. I chose to do it here to follow the
lead of Docker. Even if this is not the best decision (and I have not
opinion on that), doing the opposite would introduce confusion and
confusion rarely do good.

See: https://github.com/moby/moby/issues/10824